Nonlinear Regression in R

It performs nonlinear regression usually for pharmacokinetic and pharmacodynamic models.

nlr(Fx, Data, pNames, IE, LB, UB, Error="A", ObjFx=ObjDef, SecNames, SecForms)

Arguments

Fx
Function for structural model. It should return a vector of same length to observations.

Data
Data table which will be used in Fx. Fx should access this with e$DATA.

pNames
Parameter names in the order of Fx arguments

IE
Initial estimates of parameters

LB
Lower bound for optim function. Default value is 0.

UB
Upper bound for optim function. Default value is 1e+06.

Error
Error model. One of "A" for additive error, "POIS" for poisson error, "P" for proportional error, and "C" for combined error model.

ObjFx
Objective function to be minimized. Default is maximum likelihood estimation function.

SecNames
Names of secondary parameter estimates

SecForms
Formula to calculate the secondary parameter estimates

Details

This uses scaled transformed parameters and environment e internally.

Value

This returns point estimate, standard error, relative standard error, covariance matrix of estimates, run test results on residuals, AIC, and AICc. Details are stored in the environment e.

Examples

tData = Theoph
colnames(tData) = c("ID", "BWT", "DOSE", "TIME", "DV")

fPK = function(THETA) # Prediction function
{
  DOSE = 320000 # in microgram
  TIME = e$DATA[,"TIME"] # use data in e$DATA

  K    = THETA[1]
  Ka   = THETA[2]
  V    = THETA[3]

  P  = DOSE/V*Ka/(Ka - K) * (exp(-K*TIME) - exp(-Ka*TIME))
  return(P)
}

IDs = unique(tData[,"ID"])
nID = length(IDs)
for (i in 1:nID) {
  Data = tData[tData$ID == IDs[i],]
  Res = nlr(fPK, Data, pNames=c("k", "ka", "V"), IE=c(0.1, 3, 500),
            SecNames=c("CL", "Thalf", "MRT"), SecForms=c(~V*k, ~log(2)/k, ~1/k))
  print(paste0("## ID = ", i, "##"))
  print(Res)
}

#> [1] "## ID = 1##"
#> $Est
#>                k         ka            V  AddErrVar   AddErrSD         CL
#> PE   0.053954312  1.7774192 29394.222169  0.3896372  0.6242092 1585.94503
#> SE   0.007784005  0.2299941  1449.530064  0.1661418  0.1330819  172.85019
#> RSE 14.427031240 12.9397799     4.931343 42.6401419 21.3200709   10.89888
#>        Thalf       MRT
#> PE  12.84693 18.534200
#> SE   1.85343  2.673935
#> RSE 14.42703 14.427031
#> 
#> $Cov
#>                       k            ka             V     AddErrVar
#> k          6.059074e-05 -9.280489e-04 -9.013867e+00 -1.664726e-08
#> ka        -9.280489e-04  5.289730e-02  2.119645e+02  3.840233e-07
#> V         -9.013867e+00  2.119645e+02  2.101137e+06  3.776608e-03
#> AddErrVar -1.664726e-08  3.840233e-07  3.776608e-03  2.760311e-02
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 6
#> 
#> $run$run
#> [1] 5
#> 
#> $run$p.value
#> [1] 0.2619048
#> 
#> 
#> $AIC
#> [1] 28.84872
#> 
#> $AICc
#> [1] 35.51538
#> 
#> [1] "## ID = 2##"
#> $Est
#>              k         ka            V  AddErrVar   AddErrSD         CL
#> PE   0.1016610  1.9426639 32024.771536  0.8134795  0.9019310 3255.67032
#> SE   0.0207964  0.4307417  3064.675035  0.3468676  0.1922917  451.30587
#> RSE 20.4566171 22.1727339     9.569701 42.6399995 21.3199997   13.86215
#>         Thalf       MRT
#> PE   6.818221  9.836614
#> SE   1.394777  2.012238
#> RSE 20.456617 20.456617
#> 
#> $Cov
#>                       k            ka             V     AddErrVar
#> k          4.324903e-04 -5.234795e-03 -5.174777e+01 -2.618363e-08
#> ka        -5.234795e-03  1.855384e-01  9.264106e+02  1.125730e-07
#> V         -5.174777e+01  9.264106e+02  9.392233e+06  4.884966e-03
#> AddErrVar -2.618363e-08  1.125730e-07  4.884966e-03  1.203172e-01
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 5
#> 
#> $run$run
#> [1] 6
#> 
#> $run$p.value
#> [1] 0.6428571
#> 
#> 
#> $AIC
#> [1] 36.9459
#> 
#> $AICc
#> [1] 43.61257
#> 
#> [1] "## ID = 3##"
#> $Est
#>               k        ka            V   AddErrVar    AddErrSD         CL
#> PE  0.081424641 2.4535753 34319.365330  0.03966119  0.19915117 2794.44201
#> SE  0.003809873 0.1437188   684.951604  0.01691155  0.04245909   92.85260
#> RSE 4.679017016 5.8575265     1.995817 42.64005756 21.32002878    3.32276
#>         Thalf        MRT
#> PE  8.5127447 12.2812945
#> SE  0.3983128  0.5746439
#> RSE 4.6790170  4.6790170
#> 
#> $Cov
#>                       k            ka             V     AddErrVar
#> k          1.451513e-05 -3.094035e-04 -2.072882e+00 -2.225382e-09
#> ka        -3.094035e-04  2.065510e-02  6.848231e+01  6.715510e-08
#> V         -2.072882e+00  6.848231e+01  4.691587e+05  4.890907e-04
#> AddErrVar -2.225382e-09  6.715510e-08  4.890907e-04  2.860006e-04
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 5
#> 
#> $run$run
#> [1] 7
#> 
#> $run$p.value
#> [1] 0.3571429
#> 
#> 
#> $AIC
#> [1] 3.715465
#> 
#> $AICc
#> [1] 10.38213
#> 
#> [1] "## ID = 4##"
#> $Est
#>               k         ka            V  AddErrVar   AddErrSD         CL
#> PE   0.08746650  1.1714812 31097.469810  0.5210868  0.7218634 2719.98679
#> SE   0.01605189  0.2035763  2607.290790  0.2221923  0.1539019  328.29334
#> RSE 18.35203984 17.3776875     8.384254 42.6401720 21.3200860   12.06967
#>         Thalf       MRT
#> PE   7.924716 11.432949
#> SE   1.454347  2.098179
#> RSE 18.352040 18.352040
#> 
#> $Cov
#>                       k            ka             V     AddErrVar
#> k          2.576631e-04 -2.094472e-03 -3.555243e+01 -3.661761e-08
#> ka        -2.094472e-03  4.144333e-02  3.931963e+02  4.001436e-07
#> V         -3.555243e+01  3.931963e+02  6.797965e+06  6.916609e-03
#> AddErrVar -3.661761e-08  4.001436e-07  6.916609e-03  4.936942e-02
#> 
#> $run
#> $run$m
#> [1] 4
#> 
#> $run$n
#> [1] 6
#> 
#> $run$run
#> [1] 5
#> 
#> $run$p.value
#> [1] 0.4047619
#> 
#> 
#> $AIC
#> [1] 32.04641
#> 
#> $AICc
#> [1] 38.71308
#> 
#> [1] "## ID = 5##"
#> $Est
#>               k         ka           V  AddErrVar  AddErrSD         CL
#> PE   0.08843508  1.4715011 26925.05354  1.2239519  1.106324 2381.11936
#> SE   0.02009436  0.3192602  2739.87065  0.5218949  0.235869  366.26415
#> RSE 22.72216233 21.6962239    10.17592 42.6401459 21.320073   15.38202
#>         Thalf       MRT
#> PE   7.837921 11.307729
#> SE   1.780945  2.569361
#> RSE 22.722162 22.722162
#> 
#> $Cov
#>                       k            ka             V     AddErrVar
#> k          4.037834e-04 -3.825702e-03 -4.562695e+01 -7.384263e-08
#> ka        -3.825702e-03  1.019271e-01  6.136298e+02  1.049592e-06
#> V         -4.562695e+01  6.136298e+02  7.506891e+06  1.180134e-02
#> AddErrVar -7.384263e-08  1.049592e-06  1.180134e-02  2.723742e-01
#> 
#> $run
#> $run$m
#> [1] 4
#> 
#> $run$n
#> [1] 6
#> 
#> $run$run
#> [1] 6
#> 
#> $run$p.value
#> [1] 0.5952381
#> 
#> 
#> $AIC
#> [1] 41.43958
#> 
#> $AICc
#> [1] 48.10625
#> 
#> [1] "## ID = 6##"
#> $Est
#>               k         ka            V  AddErrVar   AddErrSD        CL
#> PE   0.09952552  1.1637350 41104.663654  0.2222050  0.4713862 4090.9629
#> SE   0.01617973  0.1850828  3235.606751  0.0947491  0.1005005  420.7268
#> RSE 16.25686601 15.9042025     7.871629 42.6403998 21.3201999   10.2843
#>         Thalf       MRT
#> PE   6.964517 10.047674
#> SE   1.132212  1.633437
#> RSE 16.256866 16.256866
#> 
#> $Cov
#>                       k            ka             V     AddErrVar
#> k          2.617837e-04 -1.988203e-03 -4.509907e+01 -3.223842e-08
#> ka        -1.988203e-03  3.425563e-02  4.527839e+02  3.985651e-07
#> V         -4.509907e+01  4.527839e+02  1.046915e+07  6.770589e-03
#> AddErrVar -3.223842e-08  3.985651e-07  6.770589e-03  8.977391e-03
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 5
#> 
#> $run$run
#> [1] 4
#> 
#> $run$p.value
#> [1] 0.1666667
#> 
#> 
#> $AIC
#> [1] 22.67088
#> 
#> $AICc
#> [1] 29.33754
#> 
#> [1] "## ID = 7##"
#> $Est
#>               k          ka            V   AddErrVar    AddErrSD          CL
#> PE   0.10224535  0.67974242 32621.575146  0.09059623  0.30099207 3335.404346
#> SE   0.01161466  0.07256341  1986.491058  0.03863041  0.06417181  207.010940
#> RSE 11.35959330 10.67513321     6.089501 42.64020001 21.32010001    6.206472
#>          Thalf       MRT
#> PE   6.7792539  9.780396
#> SE   0.7700957  1.111013
#> RSE 11.3595933 11.359593
#> 
#> $Cov
#>                       k            ka             V     AddErrVar
#> k          1.349002e-04 -6.670885e-04 -2.128023e+01 -1.436815e-08
#> ka        -6.670885e-04  5.265448e-03  1.234222e+02  8.046788e-08
#> V         -2.128023e+01  1.234222e+02  3.946147e+06  2.638606e-03
#> AddErrVar -1.436814e-08  8.046787e-08  2.638605e-03  1.492309e-03
#> 
#> $run
#> $run$m
#> [1] 4
#> 
#> $run$n
#> [1] 7
#> 
#> $run$run
#> [1] 5
#> 
#> $run$p.value
#> [1] 0.3333333
#> 
#> 
#> $AIC
#> [1] 12.80186
#> 
#> $AICc
#> [1] 19.46853
#> 
#> [1] "## ID = 8##"
#> $Est
#>               k         ka            V  AddErrVar  AddErrSD         CL
#> PE   0.09195638  1.3755270 35692.013652  0.3348497  0.578662 3282.10853
#> SE   0.01528989  0.2316089  2770.809865  0.1427802  0.123371  355.81477
#> RSE 16.62732923 16.8378316     7.763109 42.6400962 21.320048   10.84104
#>         Thalf       MRT
#> PE   7.537782 10.874721
#> SE   1.253332  1.808176
#> RSE 16.627329 16.627329
#> 
#> $Cov
#>                       k            ka             V     AddErrVar
#> k          2.337808e-04 -2.274929e-03 -3.597285e+01 -2.495078e-08
#> ka        -2.274929e-03  5.364269e-02  4.735930e+02  3.317224e-07
#> V         -3.597285e+01  4.735930e+02  7.677387e+06  5.009122e-03
#> AddErrVar -2.495078e-08  3.317224e-07  5.009122e-03  2.038620e-02
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 5
#> 
#> $run$run
#> [1] 7
#> 
#> $run$p.value
#> [1] 0.3571429
#> 
#> 
#> $AIC
#> [1] 27.18185
#> 
#> $AICc
#> [1] 33.84852
#> 
#> [1] "## ID = 9##"
#> $Est
#>                k        ka            V   AddErrVar   AddErrSD          CL
#> PE   0.086631747  8.865626 38948.216859  0.22625893  0.4756668 3374.152060
#> SE   0.009328053  2.846690  1504.712606  0.09647691  0.1014123  286.814500
#> RSE 10.767476217 32.109297     3.863367 42.64004647 21.3200232    8.500343
#>          Thalf       MRT
#> PE   8.0010759 11.543113
#> SE   0.8615139  1.242902
#> RSE 10.7674762 10.767476
#> 
#> $Cov
#>                       k            ka             V     AddErrVar
#> k          8.701257e-05 -9.311313e-03 -9.887669e+00 -7.334334e-09
#> ka        -9.311313e-03  8.103646e+00  1.962685e+03  7.025506e-07
#> V         -9.887669e+00  1.962685e+03  2.264160e+06  1.090413e-03
#> AddErrVar -7.334334e-09  7.025506e-07  1.090413e-03  9.307795e-03
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 5
#> 
#> $run$run
#> [1] 6
#> 
#> $run$p.value
#> [1] 0.6428571
#> 
#> 
#> $AIC
#> [1] 22.86985
#> 
#> $AICc
#> [1] 29.53651
#> 
#> [1] "## ID = 10##"
#> $Est
#>               k         ka           V   AddErrVar    AddErrSD          CL
#> PE  0.073965665 0.69550537 25519.76585  0.12285496  0.35050671 1887.586459
#> SE  0.006923395 0.05965454  1127.00909  0.05238562  0.07472841  108.080728
#> RSE 9.360281733 8.57715071     4.41622 42.64021517 21.32010759    5.725869
#>         Thalf       MRT
#> PE  9.3712019 13.519786
#> SE  0.8771709  1.265490
#> RSE 9.3602817  9.360282
#> 
#> $Cov
#>                       k            ka             V     AddErrVar
#> k          4.793339e-05 -3.108263e-04 -7.015432e+00 -1.223971e-08
#> ka        -3.108263e-04  3.558665e-03  5.633714e+01  9.243673e-08
#> V         -7.015432e+00  5.633714e+01  1.270149e+06  2.133120e-03
#> AddErrVar -1.223971e-08  9.243673e-08  2.133120e-03  2.744253e-03
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 6
#> 
#> $run$run
#> [1] 6
#> 
#> $run$p.value
#> [1] 0.7380952
#> 
#> 
#> $AIC
#> [1] 16.15237
#> 
#> $AICc
#> [1] 22.81904
#> 
#> [1] "## ID = 11##"
#> $Est
#>               k        ka            V   AddErrVar    AddErrSD          CL
#> PE  0.098123028 3.8490542 37945.347628  0.03874697  0.19684251 3723.312416
#> SE  0.004603223 0.2483307   768.521663  0.01652178  0.04196701  126.928376
#> RSE 4.691277227 6.4517342     2.025338 42.64019332 21.32009666    3.409018
#>         Thalf        MRT
#> PE  7.0640623 10.1912876
#> SE  0.3313947  0.4781016
#> RSE 4.6912772  4.6912772
#> 
#> $Cov
#>                       k            ka             V     AddErrVar
#> k          2.118966e-05 -5.781500e-04 -2.697293e+00 -1.807689e-09
#> ka        -5.781500e-04  6.166816e-02  1.233536e+02  7.806693e-08
#> V         -2.697293e+00  1.233536e+02  5.906255e+05  3.550742e-04
#> AddErrVar -1.807689e-09  7.806693e-08  3.550742e-04  2.729694e-04
#> 
#> $run
#> $run$m
#> [1] 4
#> 
#> $run$n
#> [1] 6
#> 
#> $run$run
#> [1] 4
#> 
#> $run$p.value
#> [1] 0.1904762
#> 
#> 
#> $AIC
#> [1] 3.458906
#> 
#> $AICc
#> [1] 10.12557
#> 
#> [1] "## ID = 12##"
#> $Est
#>               k          ka            V  AddErrVar   AddErrSD          CL
#> PE   0.10557578  0.83291039 24017.514932  0.2553549  0.5053266 2535.667942
#> SE   0.01289139  0.09971057  1557.293843  0.1088724  0.1077248  176.024973
#> RSE 12.21055245 11.97134456     6.483992 42.6356972 21.3178486    6.941957
#>          Thalf       MRT
#> PE   6.5653994  9.471869
#> SE   0.8016715  1.156568
#> RSE 12.2105525 12.210552
#> 
#> $Cov
#>                       k            ka             V    AddErrVar
#> k          1.661878e-04 -9.662099e-04 -1.812356e+01 4.339652e-09
#> ka        -9.662099e-04  9.942198e-03  1.276843e+02 4.984468e-07
#> V         -1.812356e+01  1.276843e+02  2.425164e+06 9.254575e-04
#> AddErrVar  4.339652e-09  4.984468e-07  9.254575e-04 1.185319e-02
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 5
#> 
#> $run$run
#> [1] 5
#> 
#> $run$p.value
#> [1] 0.3571429
#> 
#> 
#> $AIC
#> [1] 24.20169
#> 
#> $AICc
#> [1] 30.86835
#>