Old type WinNonlin - Least Square not MLE

It performs old type Winnonlin regression.

wnl5(Fx, Data, pNames, IE, LB, UB, Error="A", ObjFx=ObjLS)

Arguments

Fx
Function for structural model. It should return a vector of same length to observations.

Data
Data table which will be used in Fx. Fx should access this with e$DATA.

pNames
Parameter names in the order of Fx arguments

IE
Initial estimates of parameters

LB
Lower bound for optim function. Default value is 0.

UB
Upper bound for optim function. Default value is 1e+06.

Error
Error model. One of "POIS" for poisson error, "PROP" for proportional error, and others for additive error model.

ObjFx
Objective function to be minimized. Default is least square function.

Details

This uses scaled transformed parameters and environment e internally. Here we do not provide standard error. If you want standard error, use nlr.

Value

This returns point estimates, run test results on residuals, AIC, and SBC. Details are stored in the environment e.

Examples

tData = Theoph
colnames(tData) = c("ID", "BWT", "DOSE", "TIME", "DV")

fPK = function(THETA) # Prediction function
{
  DOSE = 320000 # in microgram
  TIME = e$DATA[,"TIME"] # use data in e$DATA

  K  = THETA[1]
  Ka = THETA[2]
  V  = THETA[3]
  Cp  = DOSE/V*Ka/(Ka - K)*(exp(-K*TIME) - exp(-Ka*TIME))
  return(Cp)
}

IDs = unique(tData[,"ID"])
nID = length(IDs)
for (i in 1:nID) {
  Data = tData[tData$ID == IDs[i],]
  Res = wnl5(fPK, Data, pNames=c("k", "ka", "V"), IE=c(0.1, 3, 500))
  print(paste("## ID =", i, "##"))
  print(Res)
}

#> [1] "## ID = 1 ##"
#> $PE
#>            k           ka            V 
#> 5.395422e-02 1.777418e+00 2.939423e+04 
#> 
#> $WRSS
#> [1] 4.286009
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 6
#> 
#> $run$run
#> [1] 5
#> 
#> $run$p.value
#> [1] 0.2619048
#> 
#> 
#> $AIC
#> [1] 22.00892
#> 
#> $SBC
#> [1] 23.2026
#> 
#> [1] "## ID = 2 ##"
#> $PE
#>            k           ka            V 
#> 1.016619e-01 1.942665e+00 3.202467e+04 
#> 
#> $WRSS
#> [1] 8.948304
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 5
#> 
#> $run$run
#> [1] 6
#> 
#> $run$p.value
#> [1] 0.6428571
#> 
#> 
#> $AIC
#> [1] 30.1061
#> 
#> $SBC
#> [1] 31.29979
#> 
#> [1] "## ID = 3 ##"
#> $PE
#>            k           ka            V 
#> 8.142465e-02 2.453575e+00 3.431936e+04 
#> 
#> $WRSS
#> [1] 0.4362739
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 5
#> 
#> $run$run
#> [1] 7
#> 
#> $run$p.value
#> [1] 0.3571429
#> 
#> 
#> $AIC
#> [1] -3.124334
#> 
#> $SBC
#> [1] -1.930649
#> 
#> [1] "## ID = 4 ##"
#> $PE
#>            k           ka            V 
#> 8.746660e-02 1.171479e+00 3.109748e+04 
#> 
#> $WRSS
#> [1] 5.731951
#> 
#> $run
#> $run$m
#> [1] 4
#> 
#> $run$n
#> [1] 6
#> 
#> $run$run
#> [1] 5
#> 
#> $run$p.value
#> [1] 0.4047619
#> 
#> 
#> $AIC
#> [1] 25.20661
#> 
#> $SBC
#> [1] 26.4003
#> 
#> [1] "## ID = 5 ##"
#> $PE
#>            k           ka            V 
#> 8.843543e-02 1.471499e+00 2.692502e+04 
#> 
#> $WRSS
#> [1] 13.46347
#> 
#> $run
#> $run$m
#> [1] 4
#> 
#> $run$n
#> [1] 6
#> 
#> $run$run
#> [1] 6
#> 
#> $run$p.value
#> [1] 0.5952381
#> 
#> 
#> $AIC
#> [1] 34.59978
#> 
#> $SBC
#> [1] 35.79347
#> 
#> [1] "## ID = 6 ##"
#> $PE
#>            k           ka            V 
#> 9.952616e-02 1.163725e+00 4.110450e+04 
#> 
#> $WRSS
#> [1] 2.44424
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 5
#> 
#> $run$run
#> [1] 4
#> 
#> $run$p.value
#> [1] 0.1666667
#> 
#> 
#> $AIC
#> [1] 15.83108
#> 
#> $SBC
#> [1] 17.02476
#> 
#> [1] "## ID = 7 ##"
#> $PE
#>            k           ka            V 
#> 1.022454e-01 6.797419e-01 3.262157e+04 
#> 
#> $WRSS
#> [1] 0.9965572
#> 
#> $run
#> $run$m
#> [1] 4
#> 
#> $run$n
#> [1] 7
#> 
#> $run$run
#> [1] 5
#> 
#> $run$p.value
#> [1] 0.3333333
#> 
#> 
#> $AIC
#> [1] 5.962064
#> 
#> $SBC
#> [1] 7.15575
#> 
#> [1] "## ID = 8 ##"
#> $PE
#>            k           ka            V 
#> 9.195633e-02 1.375526e+00 3.569203e+04 
#> 
#> $WRSS
#> [1] 3.683351
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 5
#> 
#> $run$run
#> [1] 7
#> 
#> $run$p.value
#> [1] 0.3571429
#> 
#> 
#> $AIC
#> [1] 20.34205
#> 
#> $SBC
#> [1] 21.53574
#> 
#> [1] "## ID = 9 ##"
#> $PE
#>            k           ka            V 
#> 8.663171e-02 8.865652e+00 3.894823e+04 
#> 
#> $WRSS
#> [1] 2.488854
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 5
#> 
#> $run$run
#> [1] 6
#> 
#> $run$p.value
#> [1] 0.6428571
#> 
#> 
#> $AIC
#> [1] 16.03005
#> 
#> $SBC
#> [1] 17.22373
#> 
#> [1] "## ID = 10 ##"
#> $PE
#>            k           ka            V 
#> 7.396572e-02 6.955052e-01 2.551976e+04 
#> 
#> $WRSS
#> [1] 1.351402
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 6
#> 
#> $run$run
#> [1] 6
#> 
#> $run$p.value
#> [1] 0.7380952
#> 
#> 
#> $AIC
#> [1] 9.31257
#> 
#> $SBC
#> [1] 10.50626
#> 
#> [1] "## ID = 11 ##"
#> $PE
#>            k           ka            V 
#> 9.812327e-02 3.849036e+00 3.794529e+04 
#> 
#> $WRSS
#> [1] 0.4262162
#> 
#> $run
#> $run$m
#> [1] 4
#> 
#> $run$n
#> [1] 6
#> 
#> $run$run
#> [1] 4
#> 
#> $run$p.value
#> [1] 0.1904762
#> 
#> 
#> $AIC
#> [1] -3.380894
#> 
#> $SBC
#> [1] -2.187208
#> 
#> [1] "## ID = 12 ##"
#> $PE
#>            k           ka            V 
#> 1.055750e-01 8.329044e-01 2.401758e+04 
#> 
#> $WRSS
#> [1] 2.809197
#> 
#> $run
#> $run$m
#> [1] 5
#> 
#> $run$n
#> [1] 5
#> 
#> $run$run
#> [1] 5
#> 
#> $run$p.value
#> [1] 0.3571429
#> 
#> 
#> $AIC
#> [1] 17.36189
#> 
#> $SBC
#> [1] 18.55557
#>